Work-in-progress revival of Thermos.
Forked off of "https://github.com/MCImaginarium/Thermos" as that seems
to be the most promising in terms of a starting point, as that seemd to
have been "working" at some point.

Note: The server jar builds, but there are errors which require manual
copying of the libs to resolve.